move `clip_grad_norm_` to utils
